requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==6.0
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-api-key==3.1.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
Faker==39.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==12.0.0
PyJWT==2.10.1
python-decouple==3.8
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
